Fix compile errors
