+++ wright/src/grammar/ast.rs
@@ -215,15 +215,20 @@ pub struct Type<'s> {
 }
 
 /// A Pattern used in pattern matching.
-#[derive(Clone, Debug)]
 #[allow(missing_docs)]
+#[derive(Clone, Debug)]
 pub enum Pattern<'s> {
-    UnderscorePattern(UnderscorePattern<'s>),
+    NumLit(NumLitPattern<'s>),
+    CharLit(CharLit<'s>),
+    StringLit(StringLit<'s>),
+    BooleanLit(BooleanLit<'s>),
+    Identifier(Identifier<'s>),
+    Underscore(Underscore<'s>),
 }
 
 /// An underscore pattern in source code.
 #[derive(Copy, Clone, Debug)]
-pub struct UnderscorePattern<'s> {
+pub struct Underscore<'s> {
     /// Associated fragment in source code.
     pub frag: Fragment<'s>,
 }
@@ -235,15 +240,4 @@ pub struct NumLitPattern<'s> {
     pub negative: bool,
     /// Inner number literal value
     pub inner: NumLit<'s>,
-}
-
-/// Pattern node for pattern matching
-#[derive(Clone, Debug)]
-pub enum Pattern<'s> {
-    NumLit(NumLitPattern<'s>),
-    CharLit(CharLit<'s>),
-    StringLit(StringLit<'s>),
-    BooleanLit(BooleanLit<'s>),
-    Identifier(Identifier<'s>),
-    Underscore(Underscore<'s>),
 }
