+++ wright/src/grammar/parsers/patterns.rs
@@ -1,15 +1,71 @@
+mod num_lit;
+
 use crate::grammar::ast::eq::AstEq;
+use crate::grammar::ast::BooleanLit;
+use crate::grammar::ast::CharLit;
+use crate::grammar::ast::Identifier;
+use crate::grammar::ast::NumLitPattern;
 use crate::grammar::ast::Pattern;
+use crate::grammar::ast::StringLit;
+use crate::grammar::ast::Underscore;
 use crate::grammar::model::{Fragment, HasFragment};
+
 use std::mem::discriminant;
 
+use nom::branch::alt;
+use nom::combinator::map;
+use nom::IResult;
+
 pub(crate) mod underscore;
+
+impl<'s> Pattern<'s> {
+    fn parse_num_lit(input: Fragment<'s>) -> IResult<Fragment<'s>, Self> {
+        map(NumLitPattern::parse, Pattern::NumLit)(input)
+    }
+
+    fn parse_char_lit(input: Fragment<'s>) -> IResult<Fragment<'s>, Self> {
+        map(CharLit::parse, Pattern::CharLit)(input)
+    }
+
+    fn parse_string_lit(input: Fragment<'s>) -> IResult<Fragment<'s>, Self> {
+        map(StringLit::parse, Pattern::StringLit)(input)
+    }
+
+    fn parse_boolean_lit(input: Fragment<'s>) -> IResult<Fragment<'s>, Self> {
+        map(BooleanLit::parse, Pattern::BooleanLit)(input)
+    }
+
+    fn parse_identifier(input: Fragment<'s>) -> IResult<Fragment<'s>, Self> {
+        map(Identifier::parse, Pattern::Identifier)(input)
+    }
+
+    fn parse_underscore(input: Fragment<'s>) -> IResult<Fragment<'s>, Self> {
+        map(Underscore::parse, Pattern::Underscore)(input)
+    }
+
+    /// Parse a pattern
+    pub fn parse(input: Fragment<'s>) -> IResult<Fragment<'s>, Self> {
+        alt((
+            Self::parse_num_lit,
+            Self::parse_char_lit,
+            Self::parse_string_lit,
+            Self::parse_boolean_lit,
+            Self::parse_identifier,
+            Self::parse_underscore,
+        ))(input)
+    }
+}
 
 impl<'s> HasFragment<'s> for Pattern<'s> {
     fn get_fragment(&self) -> Fragment<'s> {
         use Pattern::*;
         match self {
-            UnderscorePattern(p) => p.get_fragment(),
+            NumLit(p) => p.get_fragment(),
+            CharLit(p) => p.get_fragment(),
+            StringLit(p) => p.get_fragment(),
+            BooleanLit(p) => p.get_fragment(),
+            Identifier(p) => p.get_fragment(),
+            Underscore(p) => p.get_fragment(),
         }
     }
 }
@@ -27,7 +83,12 @@ impl<'s> AstEq for Pattern<'s> {
 
         use Pattern::*;
         match (fst, snd) {
-            (UnderscorePattern(a), UnderscorePattern(b)) => aeq(a, b),
+            (Underscore(a), Underscore(b)) => aeq(a, b),
+            (NumLit(a), NumLit(b)) => aeq(a, b),
+            (CharLit(a), CharLit(b)) => aeq(a, b),
+            (StringLit(a), StringLit(b)) => aeq(a, b),
+            (BooleanLit(a), BooleanLit(b)) => aeq(a, b),
+            (Identifier(a), Identifier(b)) => aeq(a, b),
             _ => unreachable!(),
         }
     }
